+++ aa_intel_tool/static/aa_intel_tool/javascript/aa-intel-tool-dscan.js
@@ -12,6 +12,8 @@ jQuery(document).ready(($) => {
     const elementShipTypesTable = $('table.aa-intel-dscan-ship-types-list');
     const elementUpwellStructuresTable = $('table.aa-intel-dscan-upwell-structures-list');
     const elementDscanCountUpwellStructures = $('span#aa-intel-dscan-upwell-structures-count');
+    const elementDeployablesTable = $('table.aa-intel-dscan-deployables-list');
+    const elementDscanCountDeployables = $('span#aa-intel-dscan-deployables-count');
 
     /**
      * Corporation info element in datatable
@@ -465,6 +467,82 @@ jQuery(document).ready(($) => {
                         const newTotal = parseInt(currentTotal) + data['count'];
 
                         elementDscanCountUpwellStructures.html(newTotal);
+
+                        $(row).attr('data-shiptype-id', data['id']);
+
+                        // Highlight
+                        $(row).mouseenter(() => {
+                            $(row).addClass('aa-intel-highlight');
+                        }).mouseleave(() => {
+                            $(row).removeClass('aa-intel-highlight');
+                        });
+                    }
+                });
+            }
+        })
+        .catch(function (error) {
+            console.log(`Error: ${error.message}`);
+        });
+
+
+    /**
+     * Datatable D-Scan Deployables on Grid
+     */
+    fetch(aaIntelToolJsOptions.ajax.getDeployablesOnGrid)
+        .then(response => {
+            if (response.ok) {
+                return Promise.resolve(response);
+            } else {
+                return Promise.reject(new Error('Failed to load'));
+            }
+        })
+        .then(response => response.json())
+        .then(tableData => {
+            $('div.aa-intel-loading-table-info-deployables').hide();
+            $('div#aa-intel-dscan-row-interesting-on-grid').show();
+
+            if (Object.keys(tableData).length === 0) {
+                $('div.aa-intel-empty-table-info-deployables').show();
+            } else {
+                $('div.col-aa-intel-deployables').show();
+
+                elementDeployablesTable.DataTable({
+                    data: tableData,
+                    paging: false,
+                    language: aaIntelToolJsL10n.dataTables.translation,
+                    lengthChange: false,
+                    dom:
+                        '<\'row\'<\'col-sm-12\'f>>' +
+                        '<\'row\'<\'col-sm-12\'tr>>' +
+                        '<\'row\'<\'col-sm-12\'i>>',
+                    columns: [
+                        {
+                            data: (data) => {
+                                return ShipInfoPanel(data);
+                            }
+                        },
+                        {
+                            data: 'count'
+                        }
+                    ],
+                    order: [
+                        [1, 'desc']
+                    ],
+                    columnDefs: [
+                        {
+                            targets: 1,
+                            width: 45,
+                            createdCell: (td) => {
+                                $(td).addClass('text-right');
+                            }
+                        }
+                    ],
+                    createdRow: (row, data) => {
+                        // Upwell Structures total count
+                        const currentTotal = elementDscanCountDeployables.html();
+                        const newTotal = parseInt(currentTotal) + data['count'];
+
+                        elementDscanCountDeployables.html(newTotal);
 
                         $(row).attr('data-shiptype-id', data['id']);